Validations in structures and other fixes.

+++ src/main/java/org/moera/lib/node/types/CommentsSliceInfo.java
@@ -11,8 +11,8 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class CommentsSliceInfo implements Cloneable {
 
-    private int before;
-    private int after;
+    private long before;
+    private long after;
     private List<CommentInfo> comments;
     private int total;
     private int totalInPast;
@@ -21,19 +21,19 @@ public class CommentsSliceInfo implements Cloneable {
     @JsonIgnore
     private Object extra;
 
-    public int getBefore() {
+    public long getBefore() {
         return before;
     }
 
-    public void setBefore(int before) {
+    public void setBefore(long before) {
         this.before = before;
     }
 
-    public int getAfter() {
+    public long getAfter() {
         return after;
     }
 
-    public void setAfter(int after) {
+    public void setAfter(long after) {
         this.after = after;
     }
